requirements.txt
aiodns==3.0.0
aiohttp==3.8.1
aiohttp-socks==0.7.1
aiosignal==1.2.0
asgiref==3.4.1
async-timeout==4.0.1
attrs==21.2.0
beautifulsoup4==4.10.0
cchardet==2.1.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click==8.0.3
Django==3.2.9
djangorestframework==3.12.4
elasticsearch==7.15.2
fake-useragent==0.1.11
frozenlist==1.2.0
geographiclib==1.52
geopy==2.2.0
googletransx==2.4.2
idna==3.3
joblib==1.1.0
multidict==5.2.0
nltk==3.6.5
numpy==1.21.4
oauthlib==3.1.1
pandas==1.3.4
pycares==4.1.2
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
python-socks==2.0.0
pytz==2021.3
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
schedule==1.1.0
scikit-learn==1.0.1
scipy==1.7.3
six==1.16.0
sklearn==0.0
soupsieve==2.3.1
sqlparse==0.4.2
threadpoolctl==3.0.0
tqdm==4.62.3
tweepy==4.4.0
twint==2.1.20
typing_extensions==4.0.0
Unidecode==1.3.2
urllib3==1.26.7
yarl==1.7.2
